Reduce the number of flags passed to compiler invoked as assembler.
Right now majority of flags passed to cc when invoked to generate
assembly code are not actuall valid in such mode. GCC does not care about
this, but clang does. It prints extra warnigns when specified flags are not
actually being used or valid in a specific compilation mode. Refoactor this
code to make both GCC and clang happy.