base-hw: Fix src/include/hw/spec/arm/lpae.h build with clang.
The code contains the following construct:

enum { X = 12; }

class A {
  public:
    static constexpr size_t Y = X;
} __attribute__((aligned(1 << Y)));

This works fine with GCC, but clang complains it has no clue what Y is
when processing the __atribute__ specifier. I'm not an expert in this
area, but it does look like clang is correct here. Technically Y should
be out of scope after class closing curly bracket.

This patch does 2 things.
1. It uses the original enum value defined outside of aligned class
   scope.
2. It uses C++11 alignas instead of algined atribute. The reason for
   this is simple. One is a compiler specific extension, while the other
   is fully defined language feature. Both should be semantically the
   same.